Add async meta cache
This might introduce an issue where we won't be able to finish song initialization in pararrel soon enough, as the fetch task will be waiting to write it self into cache. Fix by making MetaCache.set add a task to an internal queue instead. Have a thread that goes through this list and handles caching.